requirements.txt
Flask==3.1.2
pymongo==4.15.3
python-dotenv==1.1.1
python_bcrypt==0.3.2
